+++ app/evm_mempool.go
@@ -47,7 +47,9 @@ func (app *KiichainApp) configureEVMMempool(appOpts servertypes.AppOptions, logg
 			sdkmempool.NewDefaultSignerExtractionAdapter(),
 		),
 	)
-	app.SetPrepareProposal(abciProposalHandler.PrepareProposalHandler())
+	decoder := app.txConfig.TxDecoder()
+	app.SetPrepareProposal(WrapPrepareProposal(app.appCodec, decoder, abciProposalHandler.PrepareProposalHandler()))
+	app.SetProcessProposal(WrapProcessProposal(app.appCodec, decoder, abciProposalHandler.ProcessProposalHandler()))
 }
 
 // createMempoolConfig creates a new EVMMempoolConfig with the default configuration